006 Round off Grading System 07042019
Grading problem from https://www.hackerrank.com/challenges/grading/problem

If the difference between the grade and the next multiple of 5 is less than 3, round grade up to the next multiple of 5.
If the value of grade is less than 38, no rounding occurs as the result will still be a failing grade.
For example, 84 will be rounded to 85 but 29 will not be rounded because the rounding would result in a number that is less than .

Complete the function gradingStudents in the editor below. It should return an integer array consisting of rounded grades.